+++ cimd/commands/extract/coverage_xml.py
@@ -1,5 +1,7 @@
 """coverage-xml command"""
 
+from enum import Enum
+
 import click
 from defusedxml.ElementTree import ParseError, parse
 
@@ -7,9 +9,15 @@ from cimd.classes.context import ApplicationContext
 from cimd.classes.metadata import Item
 from cimd.classes.shields_link import ShieldsLink
 
-BEST_THRESHOLD = 90
-GOOD_THRESHOLD = 75
-MEDIUM_THRESHOLD = 50
+COMMAND_AND_PREFIX = "coverage-xml"
+
+
+class Threshold(float, Enum):
+    """Color thresholds"""
+
+    MEDIUM = 50
+    GOOD = 75
+    BEST = 90
 
 
 def image_for_line_rate_item(line_rate: float) -> ShieldsLink:
@@ -18,18 +26,18 @@ def image_for_line_rate_item(line_rate: float) -> ShieldsLink:
     same colors as in genbadge project:
     https://github.com/smarie/python-genbadge/blob/main/genbadge/utils_coverage.py#L105C1-L119C17
     """
-    if line_rate < MEDIUM_THRESHOLD:
+    if line_rate < Threshold.MEDIUM:
         color = "red"
-    elif line_rate < GOOD_THRESHOLD:
+    elif line_rate < Threshold.GOOD:
         color = "orange"
-    elif line_rate < BEST_THRESHOLD:
+    elif line_rate < Threshold.BEST:
         color = "green"
     else:
         color = "brightgreen"
     return ShieldsLink(label="Coverage", message=f"{line_rate!s}%", color=color)
 
 
-@click.command(name="coverage-xml")
+@click.command(name=COMMAND_AND_PREFIX)
 @click.argument(
     "XML_FILE",
     type=click.Path(exists=True, dir_okay=False, file_okay=True, readable=True, resolve_path=True),
@@ -52,7 +60,7 @@ def coverage_xml_command(app: ApplicationContext, xml_file: str, replace: bool) 
         raise click.UsageError(f"{xml_file} - {error!s}") from error
     root = xml.getroot()
     line_rate: float = round(float(root.attrib.get("line-rate")) * 100, ndigits=2)
-    key = "coverage-py-line-rate"
+    key = f"{COMMAND_AND_PREFIX}-line-rate"
     new_item = Item(
         value=str(line_rate),
         label="Coverage",